refactor inbox routes to own file
- Cleans up our user routes index.js
